+++ packages/desktop/app/api/queries/useVitals.js
@@ -40,7 +40,6 @@ export const useVitals = encounterId => {
         const configs = {
           validationCriteria: getConfigObject(id, validationCriteria),
           config: getConfigObject(id, config),
-          visualisationConfig: getConfigObject(id, dataElement.visualisationConfig),
         };
         return recordedDates.reduce(
           (state, date) => ({
