+++ packages/desktop/app/api/queries/useVitalsSurvey.js
@@ -1,4 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
+import * as yup from 'yup';
 import { useApi, isErrorUnknownAllow404s } from '../index';
 import { getConfigObject } from '../../utils';
 
@@ -10,12 +11,31 @@ export const useVitalsSurvey = () => {
 
   const { data: surveyData, isLoading } = vitalsSurvey;
 
+  const visualisationConfigSchema = yup.object().shape({
+    yAxis: yup.object().shape({
+      graphRange: yup.object().shape({
+        min: yup.number().required(),
+        max: yup.number().required(),
+      }),
+      normalRange: yup.object().shape({
+        min: yup.number().required(),
+        max: yup.number().required(),
+      }),
+      interval: yup.number().required(),
+    }),
+  });
   let visualisationConfigs = [];
   if (!isLoading && surveyData) {
-    visualisationConfigs = surveyData.components.map(({ id, dataElement }) => ({
-      key: dataElement.name,
-      ...getConfigObject(id, dataElement.visualisationConfig),
-    }));
+    visualisationConfigs = surveyData.components.map(({ id, dataElement }) => {
+      const visualisationConfigObject = getConfigObject(id, dataElement.visualisationConfig);
+      const hasVitalChart = visualisationConfigSchema.isValidSync(visualisationConfigObject);
+
+      return {
+        key: dataElement.name,
+        hasVitalChart,
+        ...getConfigObject(id, dataElement.visualisationConfig),
+      };
+    });
   }
 
   return { ...vitalsSurvey, visualisationConfigs };
